Refactor - moved searchresultitem to own component
